% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{pow}
\alias{pow}
\title{Compute power transformation}
\usage{
pow(x, a, b)
}
\arguments{
\item{x}{Vector or scalar of indepenent variable.}

\item{a}{Coefficient}

\item{b}{Exponent}
}
\value{
Values of fcn at x: \code{y}.
}
\description{
Compute power transformation
}
\examples{
pow(x=2, a=3, b=4)
pow(x=c(2,3,4), a=3, b=4)
}
